More tests. Improve CLI.
