+++ packages/cli/src/bin.ts
@@ -2,6 +2,7 @@ import CreateConfig from './commands/CreateConfig';
 import RunDriver from './commands/RunDriver';
 import RunScript from './commands/RunScript';
 import Scaffold from './commands/Scaffold';
+import createDriverCommand from './createDriverCommand';
 import { argv, parallelArgv, program, tool } from './setup';
 
 async function run() {
@@ -29,26 +30,7 @@ async function run() {
 
     // Add a command for each driver
     tool.driverRegistry.getAll().forEach((driver) => {
-      const command = new RunDriver({ driver, parallelArgv });
-      const path = driver.getName();
-
-      // Override command metadata
-      Object.assign(command.constructor, {
-        category: 'driver',
-        description:
-          driver.metadata.description || tool.msg('app:run', { title: driver.metadata.title }),
-        params: [],
-        path,
-      });
-
-      // Register sub-commands for the driver
-      driver.commands.forEach(({ path: subpath, config, runner }) => {
-        command.register<{}, []>(
-          `${path}:${subpath}`,
-          { ...config, category: 'driver' },
-          (options, params, rest) => runner(tool, options, params, rest),
-        );
-      });
+      const command = createDriverCommand(driver, parallelArgv);
 
       program.register(command);
     });
