+++ packages/cli/src/commands/BaseRunCommand.ts
@@ -1,0 +1,20 @@
+import { Arg, Command, GlobalOptions, PrimitiveType } from '@boost/cli';
+import { tool } from '../setup';
+
+export default abstract class BaseRunCommand<
+  O extends object,
+  P extends PrimitiveType[],
+  C extends object = {}
+> extends Command<GlobalOptions & O, P, C> {
+  @Arg.Number(tool.msg('app:cliOptionConcurrency'))
+  concurrency: number = 0;
+
+  @Arg.Flag(tool.msg('app:cliOptionGraph'))
+  graph: boolean = false;
+
+  @Arg.String(tool.msg('app:cliOptionStdio'), { choices: ['buffer', 'stream', 'inherit'] })
+  stdio: string = 'buffer';
+
+  @Arg.String(tool.msg('app:cliOptionWorkspaces'))
+  workspaces: string = '';
+}
