+++ packages/cli/src/commands/RunDriver.ts
@@ -1,23 +1,9 @@
-import {
-  Blueprint,
-  Driver,
-  DriverContextOptions,
-  DriverContextParams,
-  Predicates,
-} from '@beemo/core';
-import {
-  Arg,
-  Argv,
-  Command,
-  CommandMetadata,
-  Config,
-  GlobalOptions,
-  ParserOptions,
-} from '@boost/cli';
+import { Blueprint, DriverContextOptions, DriverContextParams, Predicates } from '@beemo/core';
+import { Arg, Argv, Config } from '@boost/cli';
 import { tool } from '../setup';
+import BaseRunCommand from './BaseRunCommand';
 
 export interface RunDriverConfig {
-  driver?: Driver;
   parallelArgv?: Argv[];
 }
 
@@ -26,23 +12,7 @@ export interface RunDriverConfig {
   allowVariadicParams: true,
   category: 'core',
 })
-export default class RunDriver extends Command<
-  DriverContextOptions & GlobalOptions,
-  [],
-  RunDriverConfig
-> {
-  @Arg.Number(tool.msg('app:cliOptionConcurrency'))
-  concurrency: number = 0;
-
-  @Arg.Flag(tool.msg('app:cliOptionGraph'))
-  graph: boolean = true;
-
-  @Arg.String(tool.msg('app:cliOptionStdio'), { choices: ['buffer', 'stream', 'inherit'] })
-  stdio: string = 'buffer';
-
-  @Arg.String(tool.msg('app:cliOptionWorkspaces'))
-  workspaces: string = '';
-
+export default class RunDriver extends BaseRunCommand<DriverContextOptions, [], RunDriverConfig> {
   @Arg.Params<DriverContextParams>({
     description: tool.msg('app:cliArgDriverName'),
     label: 'name',
@@ -52,34 +22,16 @@ export default class RunDriver extends Command<
   async run(name: string = '') {
     const pipeline = tool.createRunDriverPipeline(
       this.getArguments(),
-      this.options.driver?.getName() || name,
+      name,
+      this.options.parallelArgv,
     );
 
     await pipeline.run();
   }
 
-  blueprint({ instance, array, string }: Predicates): Blueprint<RunDriverConfig> {
+  blueprint({ array, string }: Predicates): Blueprint<RunDriverConfig> {
     return {
-      // @ts-expect-error Because Driver is abstract
-      driver: instance(Driver),
       parallelArgv: array(array(string())),
-    };
-  }
-
-  getParserOptions(): ParserOptions<DriverContextOptions & GlobalOptions> {
-    const { driver } = this.options;
-    const parent = super.getParserOptions();
-
-    if (!driver) {
-      return parent;
-    }
-
-    return {
-      ...parent,
-      options: {
-        ...parent.options,
-        ...driver.metadata.commandOptions,
-      },
     };
   }
 }
