+++ packages/driver-typescript/tests/commands/syncProjectRefs.test.ts
@@ -114,4 +114,293 @@ describe('syncProjectRefs()', () => {
       );
     });
   });
+
+  describe('createProjectRefConfigsInWorkspaces()', () => {
+    it('creates a source and optional test config in each package root', async () => {
+      await syncProjectRefs(tool);
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/bar/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            declarationDir: 'lib',
+            outDir: 'lib',
+            rootDir: 'src',
+          },
+          exclude: ['lib', 'tests'],
+          extends: '../../tsconfig.options.json',
+          include: ['src/**/*', 'types/**/*', '../../types/**/*'],
+          references: [{ path: '../foo' }],
+        }),
+        expect.any(Function),
+      );
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/baz/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            declarationDir: 'lib',
+            outDir: 'lib',
+            rootDir: 'src',
+          },
+          exclude: ['lib', 'tests'],
+          extends: '../../tsconfig.options.json',
+          include: ['src/**/*', 'types/**/*', '../../types/**/*'],
+          references: [{ path: '../foo' }, { path: '../bar' }],
+        }),
+        expect.any(Function),
+      );
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/baz/tests/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            composite: false,
+            emitDeclarationOnly: false,
+            noEmit: true,
+            rootDir: '.',
+          },
+          extends: '../../../tsconfig.options.json',
+          include: ['**/*', '../types/**/*', '../../../types/**/*'],
+          references: [{ path: '..' }],
+        }),
+        expect.any(Function),
+      );
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/foo/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            declarationDir: 'lib',
+            outDir: 'lib',
+            rootDir: 'src',
+          },
+          exclude: ['lib', 'tests', 'some/path'],
+          extends: '../../tsconfig.options.json',
+          include: ['src/**/*', 'types/**/*', '../../types/**/*'],
+          references: [],
+        }),
+        expect.any(Function),
+      );
+    });
+
+    it('supports emitting `declarationOnly', async () => {
+      driver.configure({
+        declarationOnly: true,
+      });
+
+      await syncProjectRefs(tool);
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/bar/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            declarationDir: 'lib',
+            outDir: 'lib',
+            rootDir: 'src',
+            emitDeclarationOnly: true,
+          },
+          exclude: ['lib', 'tests'],
+          extends: '../../tsconfig.options.json',
+          include: ['src/**/*', 'types/**/*', '../../types/**/*'],
+          references: [{ path: '../foo' }],
+        }),
+        expect.any(Function),
+      );
+    });
+
+    it('supports custom `srcFolder` and `buildFolder`', async () => {
+      driver.configure({
+        buildFolder: 'build',
+        srcFolder: 'source',
+      });
+
+      await syncProjectRefs(tool);
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/qux/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            declarationDir: 'build',
+            outDir: 'build',
+            rootDir: 'source',
+          },
+          exclude: ['build', 'tests'],
+          extends: '../../tsconfig.options.json',
+          include: ['source/**/*', 'types/**/*', '../../types/**/*'],
+          references: [],
+        }),
+        expect.anything(),
+      );
+    });
+
+    it('supports custom `typesFolder` and `testsFolder`', async () => {
+      driver.configure({
+        typesFolder: 'typings',
+        testsFolder: 'custom-tests',
+      });
+
+      await syncProjectRefs(tool);
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/foo/custom-tests/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            composite: false,
+            emitDeclarationOnly: false,
+            noEmit: true,
+            rootDir: '.',
+          },
+          extends: '../../../tsconfig.options.json',
+          include: ['**/*', '../typings/**/*', '../../../typings/**/*'],
+          references: [{ path: '..' }],
+        }),
+        expect.any(Function),
+      );
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/foo/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            declarationDir: 'lib',
+            outDir: 'lib',
+            rootDir: 'src',
+          },
+          exclude: ['lib', 'custom-tests', 'some/path'],
+          extends: '../../tsconfig.options.json',
+          include: ['src/**/*', 'typings/**/*', '../../typings/**/*'],
+          references: [],
+        }),
+        expect.any(Function),
+      );
+    });
+
+    it('excludes local types when `localTypes` is false', async () => {
+      driver.configure({ localTypes: false });
+
+      await syncProjectRefs(tool);
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/baz/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            declarationDir: 'lib',
+            outDir: 'lib',
+            rootDir: 'src',
+          },
+          exclude: ['lib', 'tests'],
+          extends: '../../tsconfig.options.json',
+          include: ['src/**/*', '../../types/**/*'],
+          references: [{ path: '../foo' }, { path: '../bar' }],
+        }),
+        expect.any(Function),
+      );
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/baz/tests/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            composite: false,
+            emitDeclarationOnly: false,
+            noEmit: true,
+            rootDir: '.',
+          },
+          extends: '../../../tsconfig.options.json',
+          include: ['**/*', '../../../types/**/*'],
+          references: [{ path: '..' }],
+        }),
+        expect.any(Function),
+      );
+    });
+
+    it('excludes global types when `globalTypes` is false', async () => {
+      driver.configure({ globalTypes: false });
+
+      await syncProjectRefs(tool);
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/baz/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            declarationDir: 'lib',
+            outDir: 'lib',
+            rootDir: 'src',
+          },
+          exclude: ['lib', 'tests'],
+          extends: '../../tsconfig.options.json',
+          include: ['src/**/*', 'types/**/*'],
+          references: [{ path: '../foo' }, { path: '../bar' }],
+        }),
+        expect.any(Function),
+      );
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/baz/tests/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            composite: false,
+            emitDeclarationOnly: false,
+            noEmit: true,
+            rootDir: '.',
+          },
+          extends: '../../../tsconfig.options.json',
+          include: ['**/*', '../types/**/*'],
+          references: [{ path: '..' }],
+        }),
+        expect.any(Function),
+      );
+    });
+
+    it('emits `onCreateProjectConfigFile` event', async () => {
+      const spy = jest.fn((filePath, config, isTests) => {
+        if (isTests) {
+          config.compilerOptions.testsOnly = true;
+        } else {
+          config.compilerOptions.srcOnly = true;
+        }
+      });
+
+      driver.onCreateProjectConfigFile.listen(spy);
+
+      await syncProjectRefs(tool);
+
+      expect(spy).toHaveBeenCalledTimes(4);
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/baz/tests/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            composite: false,
+            emitDeclarationOnly: false,
+            noEmit: true,
+            rootDir: '.',
+            // @ts-expect-error Testing purposes
+            testsOnly: true,
+          },
+          extends: '../../../tsconfig.options.json',
+          include: ['**/*', '../types/**/*', '../../../types/**/*'],
+          references: [{ path: '..' }],
+        }),
+        expect.any(Function),
+      );
+
+      expect(writeSpy).toHaveBeenCalledWith(
+        expect.stringContaining('packages/baz/tsconfig.json'),
+        driver.formatConfig({
+          compilerOptions: {
+            declarationDir: 'lib',
+            outDir: 'lib',
+            rootDir: 'src',
+            // @ts-expect-error Testing purposes
+            srcOnly: true,
+          },
+          exclude: ['lib', 'tests'],
+          extends: '../../tsconfig.options.json',
+          include: ['src/**/*', 'types/**/*', '../../types/**/*'],
+          references: [{ path: '../foo' }, { path: '../bar' }],
+        }),
+        expect.any(Function),
+      );
+    });
+  });
 });
